+++ src/Farmacia/GUICaja.java
@@ -65,8 +65,8 @@ public class GUICaja {
         productosButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                ProductoGUIDAO productoGUIDAO = new ProductoGUIDAO();
-                productoGUIDAO.main();
+                ProductoGUI productoGUI = new ProductoGUI();
+                productoGUI.main();
                 SwingUtilities.getWindowAncestor(productosButton).dispose();
 
             }
@@ -82,8 +82,8 @@ public class GUICaja {
         MOVIMIENTOSFINANCIEROSButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                MovimientosGUIDAO movimientosGUIDAO = new MovimientosGUIDAO();
-                movimientosGUIDAO.ejecutar();
+                MovimientosGUI movimientosGUI = new MovimientosGUI();
+                movimientosGUI.ejecutar();
                 SwingUtilities.getWindowAncestor(MOVIMIENTOSFINANCIEROSButton).dispose();
             }
         });
