+++ src/Farmacia/MovimientosGUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Farmacia.MovimientosGUIDAO">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Farmacia.MovimientosGUI">
   <grid id="27dc6" binding="main" layout-manager="GridLayoutManager" row-count="15" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -153,7 +153,7 @@
         </constraints>
         <properties>
           <foreground color="-14780644"/>
-          <text value="Tipo de Movimiento"/>
+          <text value="Tipo de Farmacia.M.Movimiento"/>
         </properties>
       </component>
       <component id="f2cf6" class="javax.swing.JLabel">
